requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cssselect2==0.9.0
dj-database-url==3.1.2
Django==6.0.3
fonttools==4.62.1
gunicorn==25.3.0
idna==3.11
joblib==1.5.3
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.12.0
zopfli==0.4.1
